refactor modal and take out of header component
